updated the 'connect_to_database' function to call the second menu script, adding the menu of the second menu, adding the initial body of 'create_table' function
modifying the 'connec_to_database' function to check if the second menu script is present, then running the second script using the same bash and giving it the database name as argument.
adding a simple menu in the second menu script with an body of the functions, adding the body of the function 'create_tabe' to take the name of the table as input from the user and validating it then creating 2 files (data file & metadata file) in  the diiectory of the given database.